requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
PyJWT==2.6.0
Werkzeug==2.2.3
zipp==3.15.0
